Move some methods from Traversable to Seq
