+++ src/main/java/javaslang/collection/Traversable.java
@@ -5,9 +5,7 @@
  */
 package javaslang.collection;
 
-import javaslang.FilterMonadic;
 import javaslang.Kind;
-import javaslang.Tuple;
 import javaslang.Tuple2;
 import javaslang.control.Match;
 import javaslang.control.None;
@@ -30,7 +28,6 @@ import java.util.stream.StreamSupport;
  * <li>{@link #clear()}</li>
  * <li>{@link #contains(Object)}</li>
  * <li>{@link #containsAll(Iterable)}</li>
- * <li>{@link #containsSlice(Iterable)}</li>
  * <li>{@link #head()}</li>
  * <li>{@link #headOption()}</li>
  * <li>{@link #init()}</li>
@@ -128,8 +125,6 @@ import java.util.stream.StreamSupport;
  * <ul>
  * <li>{@link #cartesianProduct()}</li>
  * <li>{@link #cartesianProduct(Iterable)}</li>
- * <li>{@link #combinations()}</li>
- * <li>{@link #combinations(int)}</li>
  * <li>{@link #distinct()}</li>
  * <li>{@link #distinct(Function)}</li>
  * <li>{@link #flatMap(Function)}</li>
@@ -236,36 +231,6 @@ public interface Traversable<T> extends TraversableOnce<T> /*,
     Traversable<T> clear();
 
     /**
-     * Returns the union of all combinations from k = 0 to length().
-     * <p>
-     * Examples:
-     * <pre>
-     * <code>
-     * [].combinations() = [[]]
-     *
-     * [1,2,3].combinations() = [
-     *   [],                  // k = 0
-     *   [1], [2], [3],       // k = 1
-     *   [1,2], [1,3], [2,3], // k = 2
-     *   [1,2,3]              // k = 3
-     * ]
-     * </code>
-     * </pre>
-     *
-     * @return the combinations of this
-     */
-    Traversable<? extends Traversable<T>> combinations();
-
-    /**
-     * Returns the k-combination of this traversable, i.e. all subset of this of k distinct elements.
-     *
-     * @param k Size of subsets
-     * @return the k-combination of this elements
-     * @see <a href="http://en.wikipedia.org/wiki/Combination">Combination</a>
-     */
-    Traversable<? extends Traversable<T>> combinations(int k);
-
-    /**
      * Tests if this Traversable contains a given value.
      *
      * @param element An Object of type A, may be null.
@@ -273,20 +238,6 @@ public interface Traversable<T> extends TraversableOnce<T> /*,
      */
     default boolean contains(T element) {
         return findFirst(e -> java.util.Objects.equals(e, element)).isDefined();
-    }
-
-    /**
-     * Tests whether this sequence contains a given sequence as a slice.
-     * <p>
-     * Note: may not terminate for infinite-sized collections.
-     *
-     * @param that the sequence to test
-     * @return true if this sequence contains a slice with the same elements as that, otherwise false.
-     * @throws NullPointerException if {@code that} is null.
-     */
-    default boolean containsSlice(Iterable<? extends T> that) {
-        Objects.requireNonNull(that, "that is null");
-        return indexOfSlice(that) >= 0;
     }
 
     /**
@@ -544,79 +495,6 @@ public interface Traversable<T> extends TraversableOnce<T> /*,
     Option<T> headOption();
 
     /**
-     * Returns the index of the first occurrence of the given element or -1 if this does not contain the given element.
-     *
-     * @param element an element
-     * @return the index of the first occurrence of the given element
-     */
-    default int indexOf(T element) {
-        return indexOf(element, 0);
-    }
-
-    /**
-     * Returns the index of the first occurrence of the given element after or at some start index
-     * or -1 if this does not contain the given element.
-     *
-     * @param element an element
-     * @param from    start index
-     * @return the index of the first occurrence of the given element
-     */
-    int indexOf(T element, int from);
-
-    /**
-     * Finds first index where this sequence contains a given sequence as a slice.
-     * <p>
-     * Note: may not terminate for infinite-sized collections.
-     *
-     * @param that the sequence to test
-     * @return the first index such that the elements of this sequence starting at this index match
-     * the elements of sequence that, or -1 of no such subsequence exists.
-     * @throws NullPointerException if {@code that} is null.
-     */
-    default int indexOfSlice(Iterable<? extends T> that) {
-        Objects.requireNonNull(that, "that is null");
-        return indexOfSlice(that, 0);
-    }
-
-    /**
-     * Finds first index after or at a start index where this sequence contains a given sequence as a slice.
-     * <p>
-     * Note: may not terminate for infinite-sized collections.
-     *
-     * @param that the sequence to test
-     * @param from the start index
-     * @return the first index &gt;= from such that the elements of this sequence starting at this index match
-     * the elements of sequence that, or -1 of no such subsequence exists.
-     * @throws NullPointerException if {@code that} is null.
-     */
-    default int indexOfSlice(Iterable<? extends T> that, int from) {
-        Objects.requireNonNull(that, "that is null");
-        class Util {
-            int indexOfSlice(Traversable<T> t, Traversable<T> slice, int from) {
-                if (t.isEmpty()) {
-                    return from == 0 && slice.isEmpty() ? 0 : -1;
-                }
-                if (from <= 0 && checkPrefix(t, slice)) {
-                    return 0;
-                }
-                int idx = indexOfSlice(t.tail(), slice, from - 1);
-                return idx >= 0 ? idx + 1 : -1;
-            }
-
-            private boolean checkPrefix(Traversable<T> t, Traversable<T> prefix) {
-                if (prefix.isEmpty()) {
-                    return true;
-                } else {
-                    return !t.isEmpty() && java.util.Objects.equals(t.head(), prefix.head())
-                            && checkPrefix(t.tail(), prefix.tail());
-                }
-            }
-        }
-
-        return new Util().indexOfSlice(this, unit(that), from);
-    }
-
-    /**
      * Dual of {@linkplain #tail()}, returning all elements except the last.
      *
      * @return a new instance containing all elements except the last.
@@ -731,101 +609,6 @@ public interface Traversable<T> extends TraversableOnce<T> /*,
             }
             return traversable.head();
         }
-    }
-
-    /**
-     * Returns the index of the last occurrence of the given element or -1 if this does not contain the given element.
-     *
-     * @param element an element
-     * @return the index of the last occurrence of the given element
-     */
-    default int lastIndexOf(T element) {
-        return lastIndexOf(element, Integer.MAX_VALUE);
-    }
-
-    /**
-     * Returns the index of the last occurrence of the given element before or at a given end index
-     * or -1 if this does not contain the given element.
-     *
-     * @param element an element
-     * @param end     the end index
-     * @return the index of the last occurrence of the given element
-     */
-    int lastIndexOf(T element, int end);
-
-    /**
-     * Finds last index where this sequence contains a given sequence as a slice.
-     * <p>
-     * Note: will not terminate for infinite-sized collections.
-     *
-     * @param that the sequence to test
-     * @return the last index such that the elements of this sequence starting a this index match the elements
-     * of sequence that, or -1 of no such subsequence exists.
-     * @throws NullPointerException if {@code that} is null.
-     */
-    default int lastIndexOfSlice(Iterable<? extends T> that) {
-        Objects.requireNonNull(that, "that is null");
-        return lastIndexOfSlice(that, Integer.MAX_VALUE);
-    }
-
-    /**
-     * Finds last index before or at a given end index where this sequence contains a given sequence as a slice.
-     *
-     * @param that the sequence to test
-     * @param end  the end index
-     * @return the last index &lt;= end such that the elements of this sequence starting at this index match
-     * the elements of sequence that, or -1 of no such subsequence exists.
-     * @throws NullPointerException if {@code that} is null.
-     */
-    default int lastIndexOfSlice(Iterable<? extends T> that, int end) {
-        Objects.requireNonNull(that, "that is null");
-        class Util {
-            int lastIndexOfSlice(Traversable<T> t, Traversable<T> slice, int end) {
-                if (end < 0) {
-                    return -1;
-                }
-                if (t.isEmpty()) {
-                    return slice.isEmpty() ? 0 : -1;
-                }
-                if (slice.isEmpty()) {
-                    int len = t.length();
-                    return len < end ? len : end;
-                }
-                Tuple2<Traversable<T>, Integer> r = findSlice(t, slice);
-                if (r == null) {
-                    return -1;
-                }
-                if (r._2 <= end) {
-                    int idx = lastIndexOfSlice(r._1.tail(), slice, end - r._2);
-                    return idx >= 0 ? idx + 1 + r._2 : r._2;
-                } else {
-                    return -1;
-                }
-
-            }
-
-            private Tuple2<Traversable<T>, Integer> findSlice(Traversable<T> t, Traversable<T> slice) {
-                if (t.isEmpty()) {
-                    return slice.isEmpty() ? Tuple.of(t, 0) : null;
-                }
-                if (checkPrefix(t, slice)) {
-                    return Tuple.of(t, 0);
-                }
-                Tuple2<Traversable<T>, Integer> idx = findSlice(t.tail(), slice);
-                return idx != null ? Tuple.of(idx._1, idx._2 + 1) : null;
-            }
-
-            private boolean checkPrefix(Traversable<T> t, Traversable<T> prefix) {
-                if (prefix.isEmpty()) {
-                    return true;
-                } else {
-                    return !t.isEmpty() && java.util.Objects.equals(t.head(), prefix.head())
-                            && checkPrefix(t.tail(), prefix.tail());
-                }
-            }
-        }
-
-        return new Util().lastIndexOfSlice(this, unit(that), end);
     }
 
     /**
