+++ src/main/java/javaslang/collection/TraversableOnce.java
@@ -109,7 +109,7 @@ public interface TraversableOnce<T> extends Iterable<T> {
     }
 
     default <K, V> Map<K, V> toMap(Function<? super T, ? extends Map.Entry<? extends K, ? extends V>> f) {
-        return HashMap.ofAll(toList().map(f::apply));
+        return HashMap.<K, V>ofAll(toList().map(f::apply));
     }
 
     /**
